Bug fix.
Correctly detects loops through back-edges, however if there was no loop detected, it would cause  a segmentation fault as the "return false" was in the wrong place.

The segmentation fault would occur as it would iterate an additional time which is greater then the node count causing the previous segmentation fault.